onlyShadeSelection option
When the onlyShadeSelection option is set to true it will only emit changes coming from the shade selection. This is nessecary when changing the selectedColor of the colorpicked depending on what it emits. Beforehand it was not possible to only change it on shading colors, the color change was also called when a mainColor was selected.

Might be wise to make it the Default for ColorChange and keep emitting changes of MainColor to MainColorChange